requirements.txt
attrs==21.2.0
bs4==0.0.1
click==8.0.1
coverage==5.5
Faker==8.9.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.6.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-cover==3.0.0
pytest-coverage==0.0
python-dateutil==2.8.1
python-dotenv==0.19.0
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
watchdog==2.1.2
Werkzeug==2.0.1
zipp==3.4.1
